+++ nuxt.config.js
@@ -129,7 +129,7 @@ export default {
   plugins: [
     '~/plugins/vue-tooltip.js',
     '~/plugins/vue-notification.js',
-    '~/plugins/compiled-markdown-directive.js',
+    '~/plugins/xss.js',
     '~/plugins/vue-syntax.js',
     '~/plugins/auth.js',
     '~/plugins/shorthands.js',
@@ -158,6 +158,7 @@ export default {
     '@nuxtjs/robots',
     '@nuxtjs/sitemap',
     '@nuxtjs/style-resources',
+    '@nuxtjs/markdownit',
     'cookie-universal-nuxt',
     '~/modules/gpt-ads',
     // The analytics module is disabled, as we are using our own solution embedded in the middleware.
@@ -209,6 +210,12 @@ export default {
   },
   styleResources: {
     scss: './assets/styles/injected.scss',
+  },
+  markdownit: {
+    runtime: true,
+    preset: 'default',
+    linkify: true,
+    breaks: true,
   },
   loading: {
     color: 'green',
