+++ package.json
@@ -16,20 +16,20 @@
     "@nuxtjs/axios": "^5.13.1",
     "@nuxtjs/dayjs": "^1.2.0",
     "@nuxtjs/google-adsense": "^1.3.0",
+    "@nuxtjs/markdownit": "^2.0.0",
     "@nuxtjs/robots": "^2.4.2",
     "@nuxtjs/sitemap": "^2.4.0",
     "@nuxtjs/style-resources": "^1.0.0",
-    "nuxt": "^2.15.3",
-    "core-js": "^3.9.1",
     "cookie-universal-nuxt": "^2.1.5",
+    "core-js": "^3.9.1",
     "highlight.js": "^10.3.2",
-    "marked": "^2.0.0",
+    "nuxt": "^2.15.3",
     "sass": "^1.32.12",
     "v-tooltip": "^2.0.3",
     "vue-click-outside": "^1.1.0",
     "vue-highlightjs": "^1.3.3",
-    "vue-notification": "^1.3.20",
     "vue-multiselect": "^2.1.6",
+    "vue-notification": "^1.3.20",
     "xss": "^1.0.8"
   },
   "devDependencies": {
