+++ pages/_type/_id/version/create.vue
@@ -1,0 +1,13 @@
+<template>
+  <div></div>
+</template>
+<script>
+export default {
+  auth: true,
+  fetch() {
+    this.$parent.mode = 'edit'
+  },
+}
+</script>
+
+<style lang="scss" scoped></style>
